Moves back the commands to step definitions per Code review comments
